Clean: according to 42 norm
