+++ src/init.c
@@ -6,13 +6,13 @@
 /*   By: mbartos <mbartos@student.42prague.com>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/25 10:22:14 by mbartos           #+#    #+#             */
-/*   Updated: 2024/01/31 11:33:31 by mbartos          ###   ########.fr       */
+/*   Updated: 2024/01/31 14:54:04 by mbartos          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "philo.h"
 
-long int	get_actual_time_ms(void)
+long int	get_actual_time(void)
 {
 	struct timeval	time;
 	long			time_in_ms;
@@ -26,7 +26,7 @@ long int	get_dinner_time(long time)
 {
 	long	party_time;
 
-	party_time = (get_actual_time_ms() - time);
+	party_time = (get_actual_time() - time);
 	return (party_time);
 }
 
@@ -40,7 +40,7 @@ void	free_t_program(t_dinner *program)
 void	init_one_philo(t_philo *philo, t_shared *shared, int id)
 {
 	philo->id = id;
-	philo->eating_start_time = get_actual_time_ms();
+	philo->eating_start_time = get_actual_time();
 	philo->hold_left_fork = 0;
 	philo->hold_right_fork = 0;
 	philo->nof_meals = 0;
@@ -70,7 +70,7 @@ void	init_fill_t_program(char **argv, t_dinner *dinner, t_shared *shared)
 	dinner->shared = shared;
 	dinner->shared->table_forks = NULL;
 	dinner->shared->forks_mutexes = NULL;
-	shared->dinner_start_time = get_actual_time_ms();
+	shared->dinner_start_time = get_actual_time();
 	dinner->shared->nof_philos = ft_int_atoi(argv[1]);
 	shared->init_time_to_die = ft_int_atoi(argv[2]);
 	shared->init_time_to_eat = ft_int_atoi(argv[3]);
